HYC-1198 - Update ApaFormatter for hyrax 3 (#883)
* Switch to apa_formatter to override, update to match hyrax 3, adjust doi inclusion to add in space when needed without having to override more methods like before

* Rubocop, add more override info

* Switch to chicago_formatter to override, update to match hyrax 3, adjust doi inclusion to add in space when needed without having to override more methods like before

* Move to correct folder

* Switch to override, update doi addition, and fix test file name
